+++ RELEASE-NOTES.md
@@ -1,4 +1,13 @@
 ## RELEASE NOTES
+### Version 2.64.30-cacheable-org-service - Sept 8 2020
+**EUI-2229** - cacheable organisation Service
+
+### Version 2.64.29-case-list-selection - September 4 2020
+**EUI-2103** - case selection
+
+### Version 2.64.28-case-list-component-reset-sel-cases - September 2 2020
+**EUI-2168** - reset selected cases on model change
+
 ### Version 2.64.27-mv-upgrade-prerelease - September 2 2020
 **EUI-2370** - fixing previous release